Key LanternInstances by full jabberId.
This preserves old behavior, and I don't think we are guaranteed
that XMPP resources (instanceIds) are guaranteed to be globally
unique.  (I don't know whether jabberIds are globally unique
forever, but a collision would be less problematic if we don't
mix up users).

Also added a bunch of overly verbose comments since this part is
prone to confusion.