Actually save counters to datastore.
Also, try and make counter handling generally more robust and
consistent by guarding write ops with transactions and invalidating,
rather than overwriting, memcache so further reads go through the
datastore.  The old scheme was more performant but it made
memcache kind of authoritative, which is bad because we can't
enforce consistency guarantees on it.