Put reading and updating of group inside the update group transaction.
I don't know that there is any point in a transaction that only
protects the datastore write and not the corresponding read (and
therefore the updates in between).

I leave the operation of collecting the memcache sharded increments
out of the transaction.  These are not idempotent and it would be
tricky to keep this correct in the face of retries.  Just leaving
any concurrent increments to be handled in the next persist cycle
maintains correctness and is much simpler.