Make Dao's ShardedCounterManager an instance member.
The CounterGroup static ShardedCounterManager was being cached in
the `group` member and never refreshed.  This was OK when we were
double-caching the counters, but upon removing that, it would
cause counters to always stay at their initially loaded value.

An option to fix this would be to intrusively invalidate the
Dao.COUNTER_MANAGER.group from the PersistController.  That would
be an ugly violation of modularity and thread unsafe (what if
group was deleted after one thread had called loadGroup() but
before it was done with the `group` member?).

This solution costs us one memcache read per instance of Dao we
create.  We'll see whether that is a performance problem.