Initialize isFallbackProxy and listenHostAndPort on fp-up notifications.
Also, initialize the user as whom the proxy is running as their own
fallbackProxyUserId.

This is for robustness in the face of proxies that can't communicate
with the controller via XMPP (in our case, because Google has stopped
accepting the refresh token we have in file for them.)  They may
still serve their primary functions (giving fallback access to
Google Accounts and Google Talk and general HTTP(S) proxying), so we
want the invites to work for users with these fallbacks.