First shot at transifexable update notification email template.
This is just a modification of the invite-notification.tmpl to
slightly change the text and add a *|VERSION|* mergeVar.

I'm not sure I was supposed to put the English version as
preliminary text in the other languages, but generate_emails.py
would crash on me if the entries were missing in the language
json files.