Stop requiring a refreshToken for inviters.
The immediate reason to do this is to avoid having to query the
refresh token of invite@getlantern.org every time we process an
invite from inviteme@getlantern.org.

In any event, we no longer need users to send us tokens on invite,
since the pool of users for which we're running fallback proxies
is more or less fixed.  Then again, I didn't want to remove that
logic altogether from controller and client because that's not
necessary for this particular feature.

What I did was stop accepting inviters with no fallbackProxyUserId,
since we should no longer have those in LanternUsers and if this
ever happens again we really want to catch it ASAP.