Ameliorate the damage of concurrent modification exceptions.
Fix NullPointerException when trying to change invite status. We'll
get a RuntimeException right after that anyway, but at least that
is more informative.

Do statistical incrementing of invites assigned to a fallback proxy
to remove some pressure from that.  In any event, avoid crashing
because of this, since invites fail to get processed in this case.